支持HTTPS

+++ build-native.sh
@@ -3,9 +3,9 @@ set -e
 MOUNT=/etc/xiaoya
 PORT1=4567
 PORT2=5344
-MEM_OPT="-Xmx512M"
+BUILD=true
 
-while getopts ":d:p:m:P:t:y" arg; do
+while getopts ":d:p:P:t:yr" arg; do
     case "${arg}" in
         d)
             MOUNT=${OPTARG}
@@ -16,8 +16,8 @@ while getopts ":d:p:m:P:t:y" arg; do
         P)
             PORT2=${OPTARG}
             ;;
-        m)
-            MEM_OPT="-Xmx${OPTARG}M"
+        r)
+            BUILD=false
             ;;
         *)
             ;;
@@ -38,31 +38,26 @@ if [ $# -gt 2 ]; then
 	PORT2=$3
 fi
 
-if [ $# -gt 3 ]; then
-	MEM_OPT="-Xmx${4}M"
-	echo "Java Memory: ${MEM_OPT}"
+if [ "$BUILD" = "true" ]; then
+  rm -rf src/main/resources/static/assets && \
+  cd web-ui && \
+  npm run build || exit 1
+  cd ../atv-cli && \
+  go build && \
+  cd .. && \
+  mvn clean package -DskipTests -Pnative || exit 1
 fi
-
-rm -rf src/main/resources/static/assets && \
-cd web-ui && \
-npm run build || exit 1
-cd ../atv-cli && \
-go build && \
-cd .. && \
-mvn clean package || exit 1
-
-cd target && java -Djarmode=layertools -jar alist-tvbox-1.0.jar extract && cd ..
-pwd
 
 echo -e "\e[36m使用配置目录：\e[0m $MOUNT"
 echo -e "\e[36m端口映射：\e[0m $PORT1:4567  $PORT2:80"
 
-docker pull haroldli/alist-base
+docker pull xiaoyaliu/alist:latest
+
 docker image prune -f
 date +%j.%H%M > data/version
-docker build -f Dockerfile-xiaoya --tag=haroldli/xiaoya-tvbox:latest . || exit 1
-docker rm -f xiaoya-tvbox xiaoya alist-tvbox 2>/dev/null
-docker run -d -p $PORT1:4567 -p $PORT2:80 -e ALIST_PORT=$PORT2 -e MEM_OPT="$MEM_OPT" -v "$MOUNT":/data --name=xiaoya-tvbox haroldli/xiaoya-tvbox:latest
+docker build -f Dockerfile-native --tag=haroldli/xiaoya-tvbox:native . || exit 1
+docker rm -f xiaoya-tvbox alist-tvbox 2>/dev/null
+docker run -d -p $PORT1:4567 -p $PORT2:80 -e ALIST_PORT=$PORT2 -v "$MOUNT":/data --name=xiaoya-tvbox haroldli/xiaoya-tvbox:native
 
 sleep 1
